% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_hexTiles.R
\name{draw_hexTiles}
\alias{draw_hexTiles}
\title{Draw hexagon tiles}
\usage{
draw_hexTiles(area, offset_x_start = 0, offset_x_end = 4,
  offset_y_start = 0, offset_y_end = 4)
}
\arguments{
\item{x}{An sp object}

\item{cellsize}{The size of the hexagons, if left blank then will take the average area of the polygons in the SpatialPolygons data.frame}
}
\value{
A SpatialPolygonsDataFrame of tessellated hexagons covering the bounding box of a SpatialPolygons
}
\description{
Draw a grid of tessellated hexagons over the bounding box of a SpatialPolygons object
}
\examples{
require(rworldmap);
data("countriesCoarseLessIslands") #  Load simple map without islands
afr <- countriesCoarseLessIslands[which(!is.na(countriesCoarseLessIslands@data$REGION) &
                                         countriesCoarseLessIslands@data$REGION=="Africa"),]
afr <- sp::spTransform(afr, CRS("+init=EPSG:32663")) # Project to equidistant grid
plot(hex_tiles(afr)[afr,]) # Clip to original shape and plot
}
