% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTilegram.R
\name{hex_tiles}
\alias{hex_tiles}
\title{Draw hexagon tiles}
\usage{
hex_tiles(x, cellsize = NULL)
}
\arguments{
\item{x}{An sp object}

\item{cellsize}{The size of the hexagons, if left blank then will take the average area of the polygons in the SpatialPolygons data.frame}
}
\value{
A SpatialPolygonsDataFrame of tessellated hexagons covering the bounding box of a SpatialPolygons
}
\description{
Draw a grid of tessellated hexagons over the bounding box of a SpatialPolygons object
}
\examples{
require(rworldmap);
data("countriesCoarseLessIslands") #  Load simple map without islands
afr <- countriesCoarseLessIslands[which(!is.na(countriesCoarseLessIslands@data$REGION) &
                                         countriesCoarseLessIslands@data$REGION=="Africa"),]
afr <- sp::spTransform(afr, CRS("+init=EPSG:32663")) # Project to equidistant grid
plot(hex_tiles(afr)[afr,]) # Clip to original shape and plot
}
