% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTilegram.R
\name{makeTilegram}
\alias{makeTilegram}
\title{Make a Tilegram}
\usage{
makeTilegram(sp)
}
\arguments{
\item{x}{An sf object}
}
\value{
A sf projected to EPSG:32663 equidistant grid
}
\description{
Make a Tilegram
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
ncTileGram <- makeTilegram(nc)


require(GADMTools);
CH_cantons <- gadm_sf_loadCountries("CHE", level=2, basefile="./")
CH_cantons <- CH_cantons$sf
CH_cantons_tilegram <- makeTilegram(CH_cantons)
}
