% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTilegram.R
\name{makeTilegram}
\alias{makeTilegram}
\title{Make a Tilegram}
\usage{
makeTilegram(sp)
}
\arguments{
\item{sp}{A SpatialPolygonDataFrame}
}
\value{
A SpatialPolygonsDataFrame projected to EPSG:32663 equidistant grid
}
\description{
Function to make a tilegram from a SpatialPolygonsDataFrame.
It draws a grid of hexagons over the bounding box of the SpatialPolygonsDataFrame and
then uses the 'Hungarian' algorithm found in the `clue` package to match hexagons to
Polygons by minimising the distance between the centre of the hexagon and the centroid of the polygon.
}
\examples{
require(rworldmap);
data("countriesCoarseLessIslands") #  Load simple map without islands
afr <- countriesCoarseLessIslands[which(!is.na(countriesCoarseLessIslands@data$REGION) &
                                         countriesCoarseLessIslands@data$REGION=="Africa"),]
tileGram <- makeTilegram(afr)
plot(tileGram)
}
